+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/systems/arm/ArmSystem.java
@@ -9,7 +9,7 @@ import com.qualcomm.robotcore.hardware.DigitalChannel;
 
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 
-import org.firstinspires.ftc.teamcode.systems.BaseSystems.System;
+import org.firstinspires.ftc.teamcode.systems.base.System;
 
 import java.util.HashSet;
