+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/systems/base/LinearEncoderSystem.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.systems.BaseSystems;
+package org.firstinspires.ftc.teamcode.systems.base;
 
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
