+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/systems/flail/Flail.java
@@ -3,7 +3,7 @@ package org.firstinspires.ftc.teamcode.systems.flail;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
 
-import org.firstinspires.ftc.teamcode.systems.BaseSystems.System;
+import org.firstinspires.ftc.teamcode.systems.base.System;
 
 public class Flail extends System {
     private DcMotor flailMotor;
